Refactor: Restructure C++ tests for better modularisation of non-shared code
Start cleanup of pytest code for C++ tests


Clean up names and remove references to test_cpp.py


WIP


Move multi-GPU code


Update doc and try un-waiving
