+++ cpp/tests/README.md
@@ -2,13 +2,19 @@
 
 This document explains how to build and run the C++ tests, and the included [resources](resources).
 
-## All-in-one script
+## Pytest Scripts
 
-The Pytest script [test_cpp.py](../../tests/integration/defs/test_cpp.py) builds TRT-LLM, builds engines, and generates expected outputs and executes the C++ tests all in one go.
+The unit tests can be launched via the Pytest script in [test_unit_tests.py](../../tests/integration/defs/cpp/test_unit_tests.py). These do not require engines to be built. The Pytest script will also build TRT-LLM.
+
+The Pytest scripts in [test_e2e.py](../../tests/integration/defs/cpp/test_e2e.py) and [test_multi_gpu.py](../../tests/integration/defs/cpp/test_multi_gpu.py) build TRT-LLM, build engines, and generate expected outputs and execute the end-to-end C++ tests all in one go.
+`test_e2e.py` and `test_multi_gpu.py` contain single and multi-device tests, respectively.
+
 To get an overview of the tests and their parameterization, call:
 
 ```bash
-pytest tests/integration/defs/test_cpp.py --collect-only
+pytest tests/integration/defs/cpp/test_unit_tests.py --collect-only
+pytest tests/integration/defs/cpp/test_e2e.py --collect-only
+pytest tests/integration/defs/cpp/test_multi_gpu.py --collect-only
 ```
 
 All tests take the number of the CUDA architecture of the GPU you wish to use as a parameter e.g. 90 for Hopper.
@@ -19,13 +25,13 @@ Example calls could look like this:
 ```bash
 export LLM_MODELS_ROOT="/path/to/model_cache"
 
-pytest tests/integration/defs/test_cpp.py::test_unit_tests[90]
+pytest tests/integration/defs/cpp/test_unit_tests.py::test_unit_tests[runtime-90]
 
-pytest tests/integration/defs/test_cpp.py::test_model[llama-90]
+pytest tests/integration/defs/cpp/test_e2e.py::test_model[llama-90]
 
-pytest tests/integration/defs/test_cpp.py::test_benchmarks[gpt-90]
+pytest tests/integration/defs/cpp/test_e2e.py::test_benchmarks[gpt-90]
 
-pytest tests/integration/defs/test_cpp.py::test_multi_gpu[90]
+pytest tests/integration/defs/cpp/test_multi_gpu.py::test_disagg[90]
 ```
 
 ## Manual steps
