+++ tests/README.md
@@ -190,11 +190,11 @@ l0_a10:
   - disaggregated/test_disaggregated.py::test_disaggregated_mixed[TinyLlama-1.1B-Chat-v1.0]
   - disaggregated/test_disaggregated.py::test_disaggregated_overlap[TinyLlama-1.1B-Chat-v1.0]
   # ------------- CPP tests ---------------
-  - test_cpp.py::test_model[medusa-86]
-  - test_cpp.py::test_model[redrafter-86]
-  - test_cpp.py::test_model[mamba-86]
-  - test_cpp.py::test_model[recurrentgemma-86]
-  - test_cpp.py::test_model[eagle-86]
+  - cpp/test_e2e.py::test_model[medusa-86]
+  - cpp/test_e2e.py::test_model[redrafter-86]
+  - cpp/test_e2e.py::test_model[mamba-86]
+  - cpp/test_e2e.py::test_model[recurrentgemma-86]
+  - cpp/test_e2e.py::test_model[eagle-86]
 ```
 
 2. Copy all items in `tests` field to a text file, for example, `a10_list.txt`. Don't forget to remove extra characters like comments and the dash marks.
@@ -204,11 +204,11 @@ disaggregated/test_disaggregated.py::test_disaggregated_single_gpu_with_mpirun[T
 disaggregated/test_disaggregated.py::test_disaggregated_cuda_graph[TinyLlama-1.1B-Chat-v1.0]
 disaggregated/test_disaggregated.py::test_disaggregated_mixed[TinyLlama-1.1B-Chat-v1.0]
 disaggregated/test_disaggregated.py::test_disaggregated_overlap[TinyLlama-1.1B-Chat-v1.0]
-test_cpp.py::test_model[medusa-86]
-test_cpp.py::test_model[redrafter-86]
-test_cpp.py::test_model[mamba-86]
-test_cpp.py::test_model[recurrentgemma-86]
-test_cpp.py::test_model[eagle-86]
+cpp/test_e2e.py::test_model[medusa-86]
+cpp/test_e2e.py::test_model[redrafter-86]
+cpp/test_e2e.py::test_model[mamba-86]
+cpp/test_e2e.py::test_model[recurrentgemma-86]
+cpp/test_e2e.py::test_model[eagle-86]
 ```
 
 3. Invoke `pytest` with TRT-LLM custom option `--test-list`:
