+++ tests/integration/defs/.test_durations
@@ -121,8 +121,8 @@
    "test_e2e.py::test_benchmark_sanity_enable_fp8[llama_7b]": 253.08591708587483,
    "test_e2e.py::test_build_time_benchmark_sanity": 165.71592589840293,
    "test_unittests.py::test_unittests_v2[unittest/trt/attention/test_bert_attention.py]": 99.96196278184652,
-   "test_cpp.py::test_benchmarks[gpt-80]": 1376.0404928650241,
-   "test_cpp.py::test_model[-gpt_session-80]": 1877.1745116069796,
+   "cpp/test_e2e.py::test_benchmarks[gpt-80]": 1376.0404928650241,
+   "cpp/test_e2e.py::test_model[-gpt_session-80]": 1877.1745116069796,
    "accuracy/test_cli_flow.py::TestPhi3Small128kInstruct::test_auto_dtype": 512.450893450994,
    "accuracy/test_llm_api.py::TestLlama3_1_8B::test_fp8_rowwise": 361.5573864541948,
    "examples/test_llama.py::test_llm_llama_v3_dora_1gpu[commonsense-llama-v3-8b-dora-r32-llama-v3-8b-hf-base_fp16]": 517.2770831151865,
@@ -154,10 +154,10 @@
    "test_unittests.py::test_unittests_v2[unittest/_torch -k \"not (modeling or multi_gpu or auto_deploy)\"]": 1186.6702785710804,
    "test_unittests.py::test_unittests_v2[unittest/_torch/modeling -k \"modeling_mixtral\"]": 208.1838396479725,
    "test_unittests.py::test_unittests_v2[unittest/_torch/multi_gpu_modeling -k \"deepseek\"]": 393.0210295501165,
-   "test_cpp.py::test_model[-gpt_executor-80]": 4016.7569622844458,
-   "test_cpp.py::test_model[-gpt_tests-80]": 1817.8153839111328,
-   "test_cpp.py::test_unit_tests[executor-80]": 339.0683519244194,
-   "test_cpp.py::test_unit_tests[kernels-80]": 846.0403860099614,
+   "cpp/test_e2e.py::test_model[-gpt_executor-80]": 4016.7569622844458,
+   "cpp/test_e2e.py::test_model[-gpt_tests-80]": 1817.8153839111328,
+   "cpp/test_unit_tests.py::test_unit_tests[executor-80]": 339.0683519244194,
+   "cpp/test_unit_tests.py::test_unit_tests[kernels-80]": 846.0403860099614,
    "test_e2e.py::test_ptp_quickstart_bert[TRTLLM-BertForSequenceClassification-bert/bert-base-uncased-yelp-polarity]": 21.019993914989755,
    "test_e2e.py::test_ptp_quickstart_bert[VANILLA-BertForSequenceClassification-bert/bert-base-uncased-yelp-polarity]": 18.753523574909195,
    "test_e2e.py::test_ptp_quickstart_multimodal[NVILA-8B-FP16-vila/NVILA-8B-video]": 278.4781197870616,
@@ -256,35 +256,35 @@
    "accuracy/test_cli_flow.py::TestTinyLlama1_1BChat::test_weight_only[int8]": 159.531545445323,
    "accuracy/test_cli_flow.py::TestTinyLlama1_1BChat::test_weight_only_int8_kv_cache[int8]": 184.35870655626059,
    "test_unittests.py::test_unittests_v2[unittest/llmapi/test_llm_models.py -m \"not (part0 or part1)\"]": 825.9972547292709,
-   "test_cpp.py::test_benchmarks[bart-90]": 271.95234084688127,
-   "test_cpp.py::test_model[-bart-90]": 391.84748707409017,
-   "test_cpp.py::test_model[-eagle-86]": 850.5158995762467,
-   "test_cpp.py::test_model[-mamba-86]": 893.8684413917363,
-   "test_cpp.py::test_model[-medusa-86]": 577.0913726426661,
-   "test_cpp.py::test_model[-redrafter-86]": 356.56682327389717,
-   "test_cpp.py::test_benchmarks[t5-90]": 244.83684724476188,
-   "test_cpp.py::test_model[-enc_dec_language_adapter-90]": 356.5558080910705,
-   "test_cpp.py::test_model[-t5-90]": 167.93334361724555,
-   "test_cpp.py::test_model[fp8-llama-90]": 810.9923318810761,
-   "test_cpp.py::test_unit_tests[batch_manager-90]": 230.49758478673175,
-   "test_cpp.py::test_unit_tests[common-90]": 12.953252204693854,
-   "test_cpp.py::test_unit_tests[executor-90]": 317.7621980938129,
-   "test_cpp.py::test_unit_tests[kernels-90]": 554.3154804841615,
-   "test_cpp.py::test_unit_tests[layers-90]": 1288.0563295381144,
-   "test_cpp.py::test_unit_tests[runtime-90]": 876.2420815587975,
-   "test_cpp.py::test_unit_tests[thop-90]": 2.2652571727521718,
-   "test_cpp.py::test_unit_tests[utils-90]": 3.6415831856429577,
+   "cpp/test_e2e.py::test_benchmarks[bart-90]": 271.95234084688127,
+   "cpp/test_e2e.py::test_model[-bart-90]": 391.84748707409017,
+   "cpp/test_e2e.py::test_model[-eagle-86]": 850.5158995762467,
+   "cpp/test_e2e.py::test_model[-mamba-86]": 893.8684413917363,
+   "cpp/test_e2e.py::test_model[-medusa-86]": 577.0913726426661,
+   "cpp/test_e2e.py::test_model[-redrafter-86]": 356.56682327389717,
+   "cpp/test_e2e.py::test_benchmarks[t5-90]": 244.83684724476188,
+   "cpp/test_e2e.py::test_model[-enc_dec_language_adapter-90]": 356.5558080910705,
+   "cpp/test_e2e.py::test_model[-t5-90]": 167.93334361724555,
+   "cpp/test_e2e.py::test_model[fp8-llama-90]": 810.9923318810761,
+   "cpp/test_unit_tests.py::test_unit_tests[batch_manager-90]": 230.49758478673175,
+   "cpp/test_unit_tests.py::test_unit_tests[common-90]": 12.953252204693854,
+   "cpp/test_unit_tests.py::test_unit_tests[executor-90]": 317.7621980938129,
+   "cpp/test_unit_tests.py::test_unit_tests[kernels-90]": 554.3154804841615,
+   "cpp/test_unit_tests.py::test_unit_tests[layers-90]": 1288.0563295381144,
+   "cpp/test_unit_tests.py::test_unit_tests[runtime-90]": 876.2420815587975,
+   "cpp/test_unit_tests.py::test_unit_tests[thop-90]": 2.2652571727521718,
+   "cpp/test_unit_tests.py::test_unit_tests[utils-90]": 3.6415831856429577,
    "accuracy/test_cli_flow.py::TestLlama3_1_8B::test_fp8_rowwise_meta_recipe": 634.7149123200215,
    "examples/test_llama.py::test_llm_llama_v2_lora_1gpu[chinese-llama-2-lora-13b-llama-v2-13b-hf-lora_fp16-base_fp16]": 895.7611340929288,
    "examples/test_nemotron_nas.py::test_nemotron_nas_summary_1gpu[DeciLM-7B]": 335.41048416192643,
    "examples/test_phi.py::test_llm_phi_lora_1gpu[Phi-3-mini-4k-instruct-ru-lora-Phi-3-mini-4k-instruct-lora_fp16-base_fp16]": 217.61977925198153,
-   "test_cpp.py::test_model[-gpt-80]": 2498.94351779297,
-   "test_cpp.py::test_unit_tests[batch_manager-80]": 380.8567730002105,
-   "test_cpp.py::test_unit_tests[common-80]": 20.237869411706924,
-   "test_cpp.py::test_unit_tests[layers-80]": 2141.2598778679967,
-   "test_cpp.py::test_unit_tests[runtime-80]": 1491.7047495394945,
-   "test_cpp.py::test_unit_tests[thop-80]": 3.3458465598523617,
-   "test_cpp.py::test_unit_tests[utils-80]": 5.461210697889328,
+   "cpp/test_e2e.py::test_model[-gpt-80]": 2498.94351779297,
+   "cpp/test_unit_tests.py::test_unit_tests[batch_manager-80]": 380.8567730002105,
+   "cpp/test_unit_tests.py::test_unit_tests[common-80]": 20.237869411706924,
+   "cpp/test_unit_tests.py::test_unit_tests[layers-80]": 2141.2598778679967,
+   "cpp/test_unit_tests.py::test_unit_tests[runtime-80]": 1491.7047495394945,
+   "cpp/test_unit_tests.py::test_unit_tests[thop-80]": 3.3458465598523617,
+   "cpp/test_unit_tests.py::test_unit_tests[utils-80]": 5.461210697889328,
    "accuracy/test_cli_flow.py::TestLlama2_7B::test_fp8_gemm_plugin": 593.3573900908232,
    "examples/test_gemma.py::test_llm_gemma_1gpu_summary_vswa[gemma-3-1b-it-other-bfloat16-8]": 195.3050664511975,
    "test_e2e.py::test_llmapi_quickstart_atexit": 110.45052940770984,
