+++ tests/integration/test_lists/test-db/l0_a30.yml
@@ -39,17 +39,17 @@ l0_a30:
       backend: cpp
   tests:
   # ------------- CPP tests ---------------
-  - test_cpp.py::test_unit_tests[batch_manager-80]
-  - test_cpp.py::test_unit_tests[common-80]
-  - test_cpp.py::test_unit_tests[executor-80]
-  - test_cpp.py::test_unit_tests[kernels-80]
-  - test_cpp.py::test_unit_tests[layers-80]
-  - test_cpp.py::test_unit_tests[runtime-80]
-  - test_cpp.py::test_unit_tests[thop-80]
-  - test_cpp.py::test_unit_tests[utils-80]
-  - test_cpp.py::test_model[-gpt-80]
-  - test_cpp.py::test_model[-gpt_executor-80]
-  - test_cpp.py::test_model[-gpt_tests-80]
+  - cpp/test_unit_tests.py::test_unit_tests[batch_manager-80]
+  - cpp/test_unit_tests.py::test_unit_tests[common-80]
+  - cpp/test_unit_tests.py::test_unit_tests[executor-80]
+  - cpp/test_unit_tests.py::test_unit_tests[kernels-80]
+  - cpp/test_unit_tests.py::test_unit_tests[layers-80]
+  - cpp/test_unit_tests.py::test_unit_tests[runtime-80]
+  - cpp/test_unit_tests.py::test_unit_tests[thop-80]
+  - cpp/test_unit_tests.py::test_unit_tests[utils-80]
+  - cpp/test_e2e.py::test_model[-gpt-80]
+  - cpp/test_e2e.py::test_model[-gpt_executor-80]
+  - cpp/test_e2e.py::test_model[-gpt_tests-80]
 - condition:
     ranges:
       system_gpu_count:
@@ -163,5 +163,5 @@ l0_a30:
       stage: post_merge
       backend: cpp
   tests:
-  - test_cpp.py::test_model[-gpt_session-80]
-  - test_cpp.py::test_benchmarks[gpt-80]
+  - cpp/test_e2e.py::test_model[-gpt_session-80]
+  - cpp/test_e2e.py::test_benchmarks[gpt-80]
